pyproject.toml
[project]
name = "ingestion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fabián Murgado", email = "fabian@fabianmurgado.com" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.77",
    "ddgs>=9.11.4",
    "fastexcel>=0.19.0",
    "langfuse>=4.0.1",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.2.2",
    "simpleeval>=1.0.7",
    "strands-agents>=1.33.0",
    "streamlit>=1.55.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "detect-secrets>=1.5.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]
